requirements.txt
# Flask核心
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0

# 数据库
psycopg2-binary==2.9.9

# 认证
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# 工具
python-dotenv==1.0.0
requests==2.31.0

# PDF生成（使用浏览器自动化，分页完美）
playwright==1.40.0

# PDF解析（用于影像报告文本提取，降级方案）
pdfplumber==0.10.3

# 注意：不再需要pdf2image和pytesseract
# 大模型（Gemini）可以直接处理PDF文件，无需转换
